+++ bindings/python/test-requirements.txt
@@ -3,3 +3,4 @@ cffi>=1.12.0,<2
 cryptography>=2
 requests_mock>=1.10
 pytest>=7.0
+unasync
